server/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
audioread==3.0.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
decorator==5.1.1
django-cors-headers==3.14.0
Django==4.2
djangorestframework==3.14.0
filelock==3.9.0
fonttools==4.39.3
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
msgpack==1.0.5
networkx==3.0
nnAudio==0.3.2
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==2.0.0
Pillow==9.3.0
pip==22.3.1
pooch==1.6.0
premailer==3.10.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.5.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
sqlparse==0.4.3
sympy==1.11.1
threadpoolctl==3.1.0
timm==0.6.13
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
typing-extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wheel==0.38.4
yagmail==0.15.293
zipp==3.15.0
